% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-state.R
\name{state<-}
\alias{state<-}
\title{Imported function}
\usage{
state(self) <- value
}
\description{
Code lifted from a internal function of madgrad package.
Set 'state' attribute of an object.
}
\author{
Daniel Falbel, \email{dfalbel@gmail.com}
}
\keyword{internal}
