% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/torchopt-package.R
\docType{package}
\name{torchopt-package}
\alias{torchopt}
\alias{torchopt-package}
\title{torchopt: R Implementation of Advanced Optimizers for Torch}
\description{
A set of R implementations of deep learning optimizers for torch.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/e-sensing/torchopt/}
}

}
\author{
\strong{Maintainer}: Gilberto Camara \email{gilberto.camara@inpe.br}

Authors:
\itemize{
  \item Rolf Simoes \email{rolf.simoes@inpe.br}
  \item Daniel Falbel \email{daniel.falbel@gmail.com}
  \item Felipe Souza \email{felipe.carvalho@inpe.br}
  \item Alber Sanchez \email{alber.ipia@inpe.br}
}

}
\keyword{internal}
